--- src/main/java/br/com/alelo/consumerpat/core/dataprovider/dao/AddressDao.java
@@ -1,6 +1,0 @@
-package br.com.alelo.consumerpat.core.dataprovider.dao;
-
-import br.com.alelo.consumerpat.core.dataprovider.entity.AddressEntity;
-
-public interface AddressDao extends BaseDao<AddressEntity> {
-}
