--- src/main/java/br/com/alelo/consumerpat/core/dataprovider/dao/ContactDao.java
@@ -1,6 +1,0 @@
-package br.com.alelo.consumerpat.core.dataprovider.dao;
-
-import br.com.alelo.consumerpat.core.dataprovider.entity.ContactEntity;
-
-public interface ContactDao extends BaseDao<ContactEntity> {
-}
