--- src/main/java/br/com/alelo/consumerpat/core/dataprovider/dao/ExtractDao.java
@@ -1,6 +1,0 @@
-package br.com.alelo.consumerpat.core.dataprovider.dao;
-
-import br.com.alelo.consumerpat.core.dataprovider.entity.ExtractEntity;
-
-public interface ExtractDao extends BaseDao<ExtractEntity> {
-}
