+++ src/main/java/br/com/alelo/consumerpat/core/mapper/domain/CardDomainMapper.java
@@ -1,6 +1,6 @@
 package br.com.alelo.consumerpat.core.mapper.domain;
 
-import br.com.alelo.consumerpat.core.dataprovider.entity.CardEntity;
+import br.com.alelo.consumerpat.dataprovider.entity.CardEntity;
 import br.com.alelo.consumerpat.core.domain.CardDomain;
 import br.com.alelo.consumerpat.core.dto.v1.request.CardV1RequestDto;
 import org.springframework.stereotype.Component;
