+++ src/main/java/br/com/alelo/consumerpat/core/mapper/response/CardV1ResponseMapper.java
@@ -1,6 +1,6 @@
 package br.com.alelo.consumerpat.core.mapper.response;
 
-import br.com.alelo.consumerpat.core.dataprovider.entity.CardEntity;
+import br.com.alelo.consumerpat.dataprovider.entity.CardEntity;
 import br.com.alelo.consumerpat.core.dto.v1.response.CardV1ResponseDto;
 
 import java.util.Set;
