--- src/main/java/br/com/alelo/consumerpat/core/usecase/AuthorizationTokenUseCase.java
@@ -1,10 +1,0 @@
-package br.com.alelo.consumerpat.core.usecase;
-
-import br.com.alelo.consumerpat.core.exception.ForbiddenException;
-
-import java.security.NoSuchAlgorithmException;
-
-public interface AuthorizationTokenUseCase {
-
-    void validateToken(String token) throws ForbiddenException, NoSuchAlgorithmException;
-}
