+++ src/main/java/br/com/alelo/consumerpat/core/usecase/impl/CardRechargeUseCaseImpl.java
@@ -1,7 +1,7 @@
 package br.com.alelo.consumerpat.core.usecase.impl;
 
-import br.com.alelo.consumerpat.core.dataprovider.dao.CardDao;
-import br.com.alelo.consumerpat.core.dataprovider.entity.CardEntity;
+import br.com.alelo.consumerpat.dataprovider.dao.CardDao;
+import br.com.alelo.consumerpat.dataprovider.entity.CardEntity;
 import br.com.alelo.consumerpat.core.domain.CardDomain;
 import br.com.alelo.consumerpat.core.exception.CardNotFoundException;
 import br.com.alelo.consumerpat.core.exception.InvalidBalanceException;
