+++ src/main/java/br/com/alelo/consumerpat/dataprovider/dao/AddressDao.java
@@ -1,0 +1,6 @@
+package br.com.alelo.consumerpat.dataprovider.dao;
+
+import br.com.alelo.consumerpat.dataprovider.entity.AddressEntity;
+
+public interface AddressDao extends BaseDao<AddressEntity> {
+}
