+++ src/main/java/br/com/alelo/consumerpat/dataprovider/dao/ContactDao.java
@@ -1,0 +1,6 @@
+package br.com.alelo.consumerpat.dataprovider.dao;
+
+import br.com.alelo.consumerpat.dataprovider.entity.ContactEntity;
+
+public interface ContactDao extends BaseDao<ContactEntity> {
+}
