+++ src/main/java/br/com/alelo/consumerpat/dataprovider/dao/ExtractDao.java
@@ -1,0 +1,6 @@
+package br.com.alelo.consumerpat.dataprovider.dao;
+
+import br.com.alelo.consumerpat.dataprovider.entity.ExtractEntity;
+
+public interface ExtractDao extends BaseDao<ExtractEntity> {
+}
