+++ src/main/java/br/com/alelo/consumerpat/dataprovider/dao/impl/AddressDaoImpl.java
@@ -1,8 +1,8 @@
-package br.com.alelo.consumerpat.dataprovider.dao;
+package br.com.alelo.consumerpat.dataprovider.dao.impl;
 
 
-import br.com.alelo.consumerpat.core.dataprovider.dao.AddressDao;
-import br.com.alelo.consumerpat.core.dataprovider.entity.AddressEntity;
+import br.com.alelo.consumerpat.dataprovider.dao.AddressDao;
+import br.com.alelo.consumerpat.dataprovider.entity.AddressEntity;
 import br.com.alelo.consumerpat.dataprovider.repository.AddressRepository;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Repository;
