+++ src/main/java/br/com/alelo/consumerpat/dataprovider/dao/impl/ContactDaoImpl.java
@@ -1,7 +1,7 @@
-package br.com.alelo.consumerpat.dataprovider.dao;
+package br.com.alelo.consumerpat.dataprovider.dao.impl;
 
-import br.com.alelo.consumerpat.core.dataprovider.dao.ContactDao;
-import br.com.alelo.consumerpat.core.dataprovider.entity.ContactEntity;
+import br.com.alelo.consumerpat.dataprovider.dao.ContactDao;
+import br.com.alelo.consumerpat.dataprovider.entity.ContactEntity;
 import br.com.alelo.consumerpat.dataprovider.repository.ContactRepository;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Repository;
