+++ src/main/java/br/com/alelo/consumerpat/dataprovider/entity/ConsumerEntity.java
@@ -1,4 +1,4 @@
-package br.com.alelo.consumerpat.core.dataprovider.entity;
+package br.com.alelo.consumerpat.dataprovider.entity;
 
 import lombok.*;
 import org.springframework.data.annotation.CreatedDate;
