+++ src/main/java/br/com/alelo/consumerpat/dataprovider/entity/ExtractEntity.java
@@ -1,4 +1,4 @@
-package br.com.alelo.consumerpat.core.dataprovider.entity;
+package br.com.alelo.consumerpat.dataprovider.entity;
 
 import br.com.alelo.consumerpat.core.enumeration.EstablishmentType;
 import lombok.AllArgsConstructor;
