+++ src/main/java/br/com/alelo/consumerpat/dataprovider/repository/AddressRepository.java
@@ -1,6 +1,6 @@
 package br.com.alelo.consumerpat.dataprovider.repository;
 
-import br.com.alelo.consumerpat.core.dataprovider.entity.AddressEntity;
+import br.com.alelo.consumerpat.dataprovider.entity.AddressEntity;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
