+++ src/main/java/br/com/alelo/consumerpat/dataprovider/repository/ContactRepository.java
@@ -1,6 +1,6 @@
 package br.com.alelo.consumerpat.dataprovider.repository;
 
-import br.com.alelo.consumerpat.core.dataprovider.entity.ContactEntity;
+import br.com.alelo.consumerpat.dataprovider.entity.ContactEntity;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
