+++ src/test/java/br/com/alelo/consumer/integration/consumer/ConsumerIntegrationTest.java
@@ -1,7 +1,7 @@
 package br.com.alelo.consumer.integration.consumer;
 
 import br.com.alelo.consumer.ConsumerPatTestApplication;
-import br.com.alelo.consumerpat.core.dataprovider.entity.ConsumerEntity;
+import br.com.alelo.consumerpat.dataprovider.entity.ConsumerEntity;
 import br.com.alelo.consumerpat.core.dto.v1.request.*;
 import br.com.alelo.consumerpat.core.dto.v1.response.ConsumerV1ResponseDto;
 import br.com.alelo.consumerpat.core.dto.v1.response.PaginatedResponseDto;
